git commit onion implementado
