+++ Proyecto.Backend/Application/Services/IAuthService.cs
@@ -1,0 +1,6 @@
+﻿using Proyecto.Backend.Domain.Entities.Models;
+
+public interface IAuthService
+{
+    LoginResponse? Login(LoginRequest request);
+}
