+++ Proyecto.Backend/Domain/Entities/Models/LoginRequest.cs
@@ -1,0 +1,8 @@
+﻿namespace Proyecto.Backend.Domain.Entities.Models
+{
+    public class LoginRequest
+    {
+        public string Usuari { get; set; } = string.Empty;
+        public string Pass { get; set; } = string.Empty;
+    }
+}
