+++ Proyecto.Backend/Infrastructure/Helpers/JwtHelper.cs
@@ -1,0 +1,39 @@
+﻿
+using System.IdentityModel.Tokens.Jwt;
+using System.Security.Claims;
+using System.Text;
+using Microsoft.IdentityModel.Tokens;
+using Proyecto.Backend.Domain.Entities.Models;
+
+public class JwtHelper
+{
+    private readonly IConfiguration _config;
+
+    public JwtHelper(IConfiguration config)
+    {
+        _config = config;
+    }
+
+    public string GenerateToken(Usuario usuario)
+    {
+        var claims = new[]
+        {
+            new Claim(ClaimTypes.Name, usuario.Usuari),
+            new Claim(ClaimTypes.Role, usuario.TipoUsuario),
+            new Claim("Ced", usuario.Ced)
+        };
+
+        var key = new SymmetricSecurityKey(Encoding.UTF8.GetBytes(_config["Jwt:Key"]));
+        var creds = new SigningCredentials(key, SecurityAlgorithms.HmacSha256);
+
+        var token = new JwtSecurityToken(
+            issuer: _config["Jwt:Issuer"],
+            audience: _config["Jwt:Audience"],
+            claims: claims,
+            expires: DateTime.Now.AddHours(1),
+            signingCredentials: creds
+        );
+
+        return new JwtSecurityTokenHandler().WriteToken(token);
+    }
+}
