--- Proyecto.Backend/Models/LoginRequest.cs
@@ -1,8 +1,0 @@
-﻿namespace Proyecto.Backend.Models
-{
-    public class LoginRequest
-    {
-        public string Usuari { get; set; } = string.Empty;
-        public string Pass { get; set; } = string.Empty;
-    }
-}
