--- Proyecto.Backend/Models/Obra.cs
@@ -1,19 +1,0 @@
-﻿using System;
-using System.Collections.Generic;
-
-namespace Proyecto.Backend.Models;
-
-public partial class Obra
-{
-    public int IdObra { get; set; }
-
-    public string Cedula { get; set; } = null!;
-
-    public string TipoObra { get; set; } = null!;
-
-    public DateOnly Fecha { get; set; }
-
-    public byte[]? Pdf { get; set; }
-
-    public virtual Usuario CedulaNavigation { get; set; } = null!;
-}
