--- Proyecto.Backend/Services/IAuthService.cs
@@ -1,6 +1,0 @@
-﻿using Proyecto.Backend.Models;
-
-public interface IAuthService
-{
-    LoginResponse? Login(LoginRequest request);
-}
